Modified Spinner incrementValue to use BigDecimal vs. BigInteger and to implement defaults vs. exception when IncrementPolicy LotSize and/or Tick are not provided ControlHelper - modified determineIncrementValue() to return BigDecimal and to accept aDefaultLotSize and aDefaultIncrementValue - modified determineIncrementValue() for INCREMENT_POLICY_LOT_SIZE to: 	- no longer throw IllegalArgumentException if aAtdl4jConfig.getInputAndFilterData().getInputIncrementPolicy_LotSize() is not set 	- to instead use new BigDecimal( aIncrement ) if aIncrement != null, otherwise to use aAtdl4jConfig.getDefaultLotSizeIncrementValue() - modified determineIncrementValue() for INCREMENT_POLICY_TICK to: 	- no longer throw IllegalArgumentException if aAtdl4jConfig.getInputAndFilterData().getInputIncrementPolicy_Tick() is not set 	- to instead use new BigDecimal( aIncrement ) if aIncrement != null, otherwise to use aAtdl4jConfig.getDefaultTickIncrementValue() - modified determineIncrementValue() for INCREMENT_POLICY_STATIC to: 	- to use aAtdl4jConfig.getDefaultIncrementValue() if aIncrement is null - modified getIncrementValue(), getInnerIncrementValue(), getOuterIncrementValue() to use BigDecimal vs. BigInteger SpinnerWidget - modified createWidget() to support ControlHelper.getIncrementValue(), ControlHelper.getInnerIncrementValue(), and ControlHelper.getOuterIncrementValue() to use BigDecimal vs. BigInteger InputAndFilterData - modified getInputIncrementPolicy_Tick(), setInputIncrementPolicy_Tick(), getInputIncrementPolicy_LotSize(), setInputIncrementPolicy_LotSize() to use BigDecimal vs. BigInteger datatype SWTAtdl4jInputAndFilterDataPanel - modofied to support InputAndFilterData inputIncrementPolicy_Tick and inputIncrementPolicy_LotSize change from BigInteger to BigDecimal Atdl4jConfig, AbstractAtdl4jConfig - added getDefaultIncrementValue(), getDefaultLotSizeIncrementValue(), getDefaultTickIncrementValue() - added setDefaultIncrementValue(), setDefaultLotSizeIncrementValue(), setDefaultTickIncrementValue()

+++ atdl4j/src/org/atdl4j/config/InputAndFilterData.java
@@ -4,6 +4,7 @@
  */
 package org.atdl4j.config;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -227,12 +228,14 @@ public class InputAndFilterData
 	
 	/**
 	 */
-	public BigInteger getInputIncrementPolicy_LotSize()
+// 4/18/2010 Scott Atwell	public BigInteger getInputIncrementPolicy_LotSize()
+	public BigDecimal getInputIncrementPolicy_LotSize()
 	{
 		String tempValue = getInputHiddenFieldValue( INPUT_FIELD_NAME_INCREMENT_POLICY_LotSize );
 		if ( tempValue != null )
 		{
-			return new BigInteger( tempValue );
+// 4/18/2010 Scott Atwell			return new BigInteger( tempValue );
+			return new BigDecimal( tempValue );
 		}
 		else
 		{
@@ -242,12 +245,14 @@ public class InputAndFilterData
 	
 	/**
 	 */
-	public BigInteger getInputIncrementPolicy_Tick()
+// 4/18/2010 Scott Atwell	public BigInteger getInputIncrementPolicy_Tick()
+	public BigDecimal getInputIncrementPolicy_Tick()
 	{
 		String tempValue = getInputHiddenFieldValue( INPUT_FIELD_NAME_INCREMENT_POLICY_Tick );
 		if ( tempValue != null )
 		{
-			return new BigInteger( tempValue );
+// 4/18/2010 Scott Atwell			return new BigInteger( tempValue );
+			return new BigDecimal( tempValue );
 		}
 		else
 		{
@@ -257,7 +262,8 @@ public class InputAndFilterData
 
 	/**
 	 */
-	public void setInputIncrementPolicy_LotSize( BigInteger aLotSize )
+// 4/18/2010 Scott Atwell	public void setInputIncrementPolicy_LotSize( BigInteger aLotSize )
+	public void setInputIncrementPolicy_LotSize( BigDecimal aLotSize )
 	{
 		if ( aLotSize != null )
 		{
@@ -271,7 +277,8 @@ public class InputAndFilterData
 	
 	/**
 	 */
-	public void setInputIncrementPolicy_Tick( BigInteger aTick )
+// 4/18/2010 Scott Atwell	public void setInputIncrementPolicy_Tick( BigInteger aTick )
+	public void setInputIncrementPolicy_Tick( BigDecimal aTick )
 	{
 		if ( aTick != null )
 		{
